+++ waspc/data/Generator/templates/sdk/wasp/auth/utils.ts
@@ -40,6 +40,7 @@ export type OAuthProviderData = {}
 export type PossibleProviderData = {
   email: EmailProviderData;
   username: UsernameProviderData;
+  discord: OAuthProviderData;
   google: OAuthProviderData;
   keycloak: OAuthProviderData;
   github: OAuthProviderData;