+++ web/src/components/AuthMethodsGrid.tsx
@@ -27,7 +27,12 @@ export function AuthMethodsGrid() {
       title: 'Keycloak',
       description: 'Users sign in with their Keycloak account',
       linkToDocs: '/docs/auth/social-auth/keycloak',
-    }
+    },
+    {
+      title: 'Discord',
+      description: 'Users sign in with their Discord account',
+      linkToDocs: '/docs/auth/social-auth/discord',
+    },
   ]
   return (
     <>